# artifacts/api-server/src/routes/reports.ts
import { Router } from "express";
import { getAuth } from "@clerk/express";
import { db, reportsTable, decisionsTable, usersTable } from "@workspace/db";
import { eq, and, inArray } from "drizzle-orm";
import { GenerateReportBody } from "@workspace/api-zod";

const router = Router();

const requireAuth = (req: any, res: any, next: any) => {
  const auth = getAuth(req);
  const userId = auth?.userId;
  if (!userId) return res.status(401).json({ error: "Unauthorized" });
  req.userId = userId;
  db.insert(usersTable)
    .values({ id: userId, email: (auth.sessionClaims?.email as string) ?? "" })
    .onConflictDoNothing()
    .execute()
    .catch(() => {});
  next();
};

function generateMarkdownReport(decisions: any[], reportType: string): string {
  const typeName =
    reportType === "board_briefing"
      ? "Board Briefing"
      : reportType === "weekly"
        ? "Weekly Intelligence Digest"
        : "Pattern Analysis";

  const date = new Date().toLocaleDateString("en-US", {
    year: "numeric",
    month: "long",
    day: "numeric",
  });

  let md = `# ${typeName}\n**Generated:** ${date}\n\n---\n\n`;
  md += `## Executive Summary\n\nThis report covers **${decisions.length} decision${decisions.length === 1 ? "" : "s"}** made during the selected period.\n\n`;

  md += `## Decision Log\n\n`;
  decisions.forEach((d, i) => {
    md += `### ${i + 1}. ${d.title}\n`;
    md += `- **Stakes:** ${d.stakes.toUpperCase()}\n`;
    md += `- **Platform:** ${d.sourcePlatform ?? "Manual"}\n`;
    md += `- **Status:** ${d.status}\n`;
    if (d.description) md += `- **Context:** ${d.description}\n`;
    md += `\n`;
  });

  const highStakes = decisions.filter((d) => d.stakes === "high" || d.stakes === "critical");
  if (highStakes.length > 0) {
    md += `## High-Stakes Decisions Requiring Attention\n\n`;
    highStakes.forEach((d) => {
      md += `- **${d.title}** (${d.stakes.toUpperCase()})\n`;
    });
    md += "\n";
  }

  md += `## Recommendations\n\n`;
  md += `1. Schedule outcome reviews for all decisions logged more than 90 days ago.\n`;
  md += `2. Review patterns and alerts to identify recurring decision biases.\n`;
  md += `3. Ensure external data sources are being referenced for high-stakes decisions.\n\n`;

  md += `---\n*Generated by Decision Brain Intelligence Engine*`;
  return md;
}

router.get("/", requireAuth, async (req: any, res) => {
  try {
    const reports = await db
      .select()
      .from(reportsTable)
      .where(eq(reportsTable.userId, req.userId));
    return res.json(reports);
  } catch (err) {
    req.log.error({ err }, "Failed to list reports");
    return res.status(500).json({ error: "Internal server error" });
  }
});

router.post("/generate", requireAuth, async (req: any, res) => {
  try {
    const parsed = GenerateReportBody.safeParse(req.body);
    if (!parsed.success) return res.status(400).json({ error: parsed.error.message });
    const { decisionIds, reportType = "board_briefing", periodStart, periodEnd } = parsed.data;

    const decisions = await db
      .select()
      .from(decisionsTable)
      .where(and(eq(decisionsTable.userId, req.userId), inArray(decisionsTable.id, decisionIds)));

    const content = generateMarkdownReport(decisions, reportType);
    const title = `${reportType === "board_briefing" ? "Board Briefing" : reportType === "weekly" ? "Weekly Digest" : "Pattern Analysis"} — ${new Date().toLocaleDateString()}`;

    const [report] = await db
      .insert(reportsTable)
      .values({
        userId: req.userId,
        title,
        reportType,
        content,
        decisionIds,
        periodStart: periodStart ? new Date(periodStart) : undefined,
        periodEnd: periodEnd ? new Date(periodEnd) : undefined,
      })
      .returning();

    return res.status(201).json(report);
  } catch (err) {
    req.log.error({ err }, "Failed to generate report");
    return res.status(500).json({ error: "Internal server error" });
  }
});

router.get("/:id", requireAuth, async (req: any, res) => {
  try {
    const [report] = await db
      .select()
      .from(reportsTable)
      .where(and(eq(reportsTable.id, req.params.id), eq(reportsTable.userId, req.userId)));
    if (!report) return res.status(404).json({ error: "Not found" });
    return res.json(report);
  } catch (err) {
    req.log.error({ err }, "Failed to get report");
    return res.status(500).json({ error: "Internal server error" });
  }
});

export default router;