# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Decision Brain API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List board reports
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a board report
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateReportBody"
      responses:
        "201":
          description: Report generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a report
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Report:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
          nullable: true
        reportType:
          type: string
          enum: [board_briefing, weekly, pattern_analysis]
        content:
          type: string
          nullable: true
        periodStart:
          type: string
          format: date
          nullable: true
        periodEnd:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - reportType
        - createdAt
    GenerateReportBody:
      type: object
      properties:
        decisionIds:
          type: array
          items:
            type: string
        reportType:
          type: string
          enum: [board_briefing, weekly, pattern_analysis]
          default: board_briefing
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
      required:
        - decisionIds
...